requirements.txt
# Core dependencies
pynput>=1.7.6
sounddevice>=0.4.6
numpy>=1.24.0
pyyaml>=6.0

# TTS (Kokoro via mlx-audio, Apple Silicon optimized)
mlx-audio
soundfile
misaki<0.8
num2words
phonemizer
spacy
espeakng-loader

# Overlay indicator (macOS, PyObjC)
pyobjc-framework-Cocoa
pyobjc-framework-Quartz

# Speech-to-text (choose one backend)
# For CPU (cross-platform):
faster-whisper>=0.10.0

# For Apple Silicon (faster, install separately):
# pip install mlx-whisper

# AFK mode (Telegram)
requests>=2.28.0
